+++ R/get_vcov.R
@@ -1,11 +1,11 @@
 #' Get a named variance-covariance matrix from a model object (internal function)
-#' 
+#'
 #' @inheritParams marginaleffects
 #' @return A named square matrix of variance and covariances. The names must match the coefficient names.
 #' @rdname get_vcov
 #' @keywords internal
 #' @export
-get_vcov <- function (model, ...) {
+get_vcov <- function(model, ...) {
     UseMethod("get_vcov", model)
 }
 
@@ -15,5 +15,9 @@ get_vcov <- function (model, ...) {
 get_vcov.default <- function(model, ...) {
     # suppress "Re-fitting to get Hessian"
     out <- suppressMessages(stats::vcov(model))
+    # survival::coxph with 1 regressor produces a vector
+    if (!isTRUE(checkmate::check_matrix(out))) {
+        out <- insight::get_varcov(model)
+    }
     return(out)
 }