+++ R/plot_cap.R
@@ -20,12 +20,10 @@
 #'
 plot_cap <- function(model, 
                      condition,
-                     # type = "response",
+                     type = "response",
                      conf.int = TRUE,
                      conf.level = 0.95,
                      draw = TRUE) {
-
-    # TODO: reinstate the `type` argument when insight supports it
 
     # get data to know over what range of values we should plot
     dat <- insight::get_data(model)
@@ -84,10 +82,11 @@ plot_cap <- function(model,
     # create data
     at_list[["model"]] = model
     nd <- do.call("typical", at_list)
-    datplot <- predictions(model, 
-                             newdata = nd, 
-                             conf.int = conf.int,
-                             conf.level = conf.level)
+    datplot <- predictions(model,
+                           newdata = nd,
+                           type = type,
+                           conf.int = conf.int,
+                           conf.level = conf.level)
     colnames(datplot)[colnames(datplot) == condition1] <- "condition1"
     colnames(datplot)[colnames(datplot) == condition2] <- "condition2"
     colnames(datplot)[colnames(datplot) == condition3] <- "condition3"