+++ tests/testthat/test-pkg-bife.R
@@ -1,8 +1,8 @@
 skip_if_not_installed("bife")
+requiet("bife")
 
-test_that("bife: no validity check", {
-    requiet("bife")
-    dataset <- psid
+test_that("marginaleffects: bife: no validity", {
+    dataset <- bife::psid
     mod <- bife(LFP ~ AGE + I(INCH / 1000) + KID1 + KID2 + KID3 | ID, data = dataset)
     mfx <- marginaleffects(mod)
     tid <- tidy(mfx)
@@ -10,4 +10,11 @@ test_that("bife: no validity check", {
     expect_true("std.error" %in% colnames(tid))
     expect_false(any(tid$estimate == 0))
     expect_false(any(tid$std.error == 0))
+})
+
+test_that("predictions: bife: no validity", {
+    dataset <- bife::psid
+    mod <- bife(LFP ~ AGE + I(INCH / 1000) + KID1 + KID2 + KID3 | ID, data = dataset)
+    pred <- predictions(mod)
+    expect_predictions(pred, n_row = 1, se = FALSE)
 })