+++ NEWS.md
@@ -1,6 +1,35 @@
+# marginaleffects 0.2.0
+
+Breaking change:
+
+* `data` argument becomes `newdata` in all functions.
+
+`marginaleffects`:
+
+* Support `lme4:glmer.nb`
+* Support `mgcv::gam`
+* Support `ordinal::clm`
+* Support `mgcv`
+
+`marginalmeans`:
+
+* New `variables_grid` argument
+
+`predictions`:
+
+* Support `mgcv`
+
+`plot_cap`
+
+* New `type` argument
+
+Misc:
+
+* New validity checks and tests
+
 # marginaleffects 0.1.0
 
-First release. Yay!
+First release. Bravo!
 
 Thanks to Marco Avina Mendoza, Resul Umit, and all those who offered comments
 and suggestions.